requirements.txt
google-cloud-vision~=3.10.2
protobuf~=6.31.1
symspellpy~=6.9.0
wordfreq~=3.1.1
Pillow~=11.3.0
ipython